Safari-611.1.10/LayoutTests/js/dfg-dead-unreachable-code-with-chain-of-dead-unchecked-nodes-expected.txt

Tests that code that is dead, unreachable, and contains a chain of nodes that use each other in an untyped way doesn't result in the IR getting corrupted.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(42)(false) is 5 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
